==> 3_climate_bogwell.Rmd <==
---
title: 'Marcell Experimental Forest: S2 Climate and Bogwell Data'
author: "Mia Forsline"
date: '`r format(Sys.time(), "%B %d, %Y")`'
knit: (function(inputFile, encoding) { 
      rmarkdown::render(inputFile,
                        encoding=encoding, output_file="/Users/miaforsline/Desktop/climate_bogwell/index.html") })
output:
  html_document:
    theme: flatly
    code_folding: hide
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
editor_options:
  markdown:
    wrap: 72
---

# Set Up

Install/load necessary R packages

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      message = FALSE, 
                      warning = FALSE)

#install packages if necessary, then load libraries
if (!require(librarian)){
  install.packages("librarian")
  library(librarian)
}

librarian::shelf(
  here, #for file paths 
  janitor, 
  kableExtra,
  plotly, #for creating interactive plots 
  readxl, 
  tidyverse,
  zoo #for interpolating missing data points 
  )
```

Set date range

```{r min_max_year}
min_year = 2017
max_year = 2021
```

Create sub-directories if necessary

```{r directories}
output_dir <- file.path(here("intermediate_data"))

if (!dir.exists(output_dir)){
dir.create(output_dir)
} else {
    print("intermediate_data directory already exists!")
}

output_dir <- file.path(here("figures"))

if (!dir.exists(output_dir)){
dir.create(output_dir)
} else {
    print("figures directory already exists!")
}
```

# Climate Data

# Precipitation

More information about precipitation data collected at the Marcell
Experimental Forest can be found on the [Environmental Data Initiative
Repository
Portal](https://portal.edirepository.org/nis/mapbrowse?scope=edi&identifier=563).

## Cumulative Precipitation Every 15min

-   Read in the data from the following Box file path:
    `External-MEF_DATA/EDI/precipitation_15min/edi.849.2/data_objects/`

-   Clean the data into a tidy format

    -   Subset precipitation data to 2017 - present

-   Plot

```{r precip_15}
#set file path to read in data from Box 
filepath = "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/EDI/precipitation_15min/edi.849.2/data_objects/"

#read in the raw data 
precip_15 <- read_csv(here(filepath, "MEF_South_precip_15min.csv"))

#clean the data 
precip_15_clean <- precip_15 %>% 
  clean_names() %>% 
  mutate(year = format(as.POSIXct(timestamp, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y'),
         year = as.numeric(year)) %>% 
  subset(year >= min_year) %>% 
  rename(precip_in = south_pcp) %>% 
  mutate(precip_10 = (precip_in / 10)) %>% 
  pivot_longer(
    cols = c("precip_in", "precip_10"),
    names_to = "precip_type",
    values_to = "precip_in"
  ) %>% 
  rename(datetime = timestamp)

#save clean data CSV to use in future RMD files
write.csv(x = precip_15_clean, 
          file = file.path(here("intermediate_data", "precip_15_clean.csv")),
          row.names = FALSE)

#plot 
p_15 <- ggplot(data = precip_15_clean) + 
  geom_line(aes(x = datetime, 
                y = precip_in, 
                color = precip_type),
            alpha = 0.5) + 
  theme_classic() + 
  labs(x = "Time", 
       y = "Precipitation (cm)", 
       title = paste0("Precipitation Sum Every 15 Min (", min_year, "-", max_year, ")"),
       color = "Precipitation Type") + 
  scale_color_discrete(labels = c("inches / 10", "inches"))

#static plot
p_15

#interactive plot 
#ggplotly(p_15)
```

## Cumulative Precipitation Every 6 Hours

```{r precip_6h}
#read in 15min precip data 
precip_6h <- precip_15 

#clean and sum precip by 6 hours 
precip_6h_clean <- precip_6h %>% 
  clean_names() %>% 
  mutate(timestamp = lubridate::floor_date(timestamp, unit = "6 hours")) %>%
    group_by(timestamp) %>%
    dplyr::summarize(south_pcp = sum(south_pcp)) %>%
    ungroup() %>% 
  rename(datetime = timestamp,
         precip_in = south_pcp) %>% 
  mutate(
    year = format(as.POSIXct(datetime, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y'),
         year = as.numeric(year),
     date = format(as.POSIXct(datetime, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y-%m-%d'),
    precip_10_in = precip_in / 10
  ) %>% 
  subset(year >= min_year & year <= max_year)

#save clean data CSV to use in future RMD files
write.csv(x = precip_6h_clean, 
          file = file.path(here("intermediate_data", "precip_6h_clean.csv")),
          row.names = FALSE)

#plot
p_precip_6h <- ggplot(data = precip_6h_clean) + 
  geom_col(aes(x = datetime, y = precip_in)) + 
  theme_classic() + 
  labs(x = "Time",
       y = "Cumulative Precipitation (in)", 
       title = paste0("Precipitation Sum Every 6 Hours (", min_year, "-", max_year, ")")) + 
  theme(plot.title = element_text(hjust = 0.5),
          plot.subtitle = element_text(hjust = 0.5))

#static plot
p_precip_6h

#interactive plot
#ggplotly(p_precip_6h)

```

## Cumulative Precipitation Every 12 Hours

```{r precip_12h}
#read in 15min precip data 
precip_12h <- precip_15 

#clean and sum precip by 6 hours 
precip_12h_clean <- precip_12h %>% 
  clean_names() %>% 
  mutate(timestamp = lubridate::floor_date(timestamp, unit = "12 hours")) %>%
    group_by(timestamp) %>%
    dplyr::summarize(south_pcp = sum(south_pcp)) %>%
    ungroup() %>% 
  rename(datetime = timestamp,
         precip_in = south_pcp) %>% 
  mutate(
    year = format(as.POSIXct(datetime, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y'),
         year = as.numeric(year),
    date = format(as.POSIXct(datetime, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y-%m-%d'),
    precip_10_in = precip_in / 10
  ) %>% 
  subset(year >= min_year & year <= max_year)

#save clean data CSV to use in future RMD files
write.csv(x = precip_12h_clean, 
          file = file.path(here("intermediate_data", "precip_12h_clean.csv")),
          row.names = FALSE)

#plot
p_precip_12h <- ggplot(data = precip_12h_clean) + 
  geom_col(aes(x = datetime, y = precip_in)) + 
  theme_classic() + 
  labs(x = "Time",
       y = "Cumulative Precipitation (in)", 
       title = paste0("Precipitation Sum Every 12 Hours (", min_year, "-", max_year, ")")) + 
  theme(plot.title = element_text(hjust = 0.5),
          plot.subtitle = element_text(hjust = 0.5))

#static plot 
p_precip_12h

#interactive plot
#ggplotly(p_precip_12h)
```

## Cumulative Precipitation Daily

-   Read in the data from the following Box file path:
    `External-MEF_DATA/EDI/precipitation_daily/edi.563.3/data_objects/`

-   Clean the data into a tidy format

    -   Subset precipitation data to 2017 - present

-   Plot

```{r precip_daily}
#set file path to read in data from Box 
filepath = "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/EDI/precipitation_daily/edi.563.3/data_objects/"

#read in the raw data 
precip_daily <- read_csv(here(filepath, "MEF_precipitation_daily.csv"))

#clean the data 
precip_daily_clean <- precip_daily %>% 
  clean_names() %>% 
  select(date, south_pcp, south_flag) %>% 
  mutate(year = format(as.POSIXct(date, format = '%Y-%m-%d', 
                                  tz = "GMT"), 
                       format = '%Y'),
         year = as.numeric(year)) %>% 
  subset(year >= min_year) %>% 
  rename(precip_in = south_pcp) %>% 
  mutate(precip_10 = (precip_in / 10)) %>% 
  pivot_longer(
    cols = c("precip_in", "precip_10"),
    names_to = "precip_type",
    values_to = "precip_in"
  )

#save clean data CSV to use in future RMD files
write.csv(x = precip_daily_clean, 
          file = file.path(here("intermediate_data", "precip_daily_clean.csv")),
          row.names = FALSE)

#plot 
p_daily <- ggplot(data = precip_daily_clean) + 
  geom_line(aes(x = date, 
                y = precip_in, 
                color = precip_type),
            alpha = 0.5) + 
  theme_classic() + 
  labs(x = "Time", 
       y = "South Precipitation (in)", 
       title = paste0("Daily Cumulative Precipitation (", min_year, "-", max_year, ")"),
       color = "Precipitation Type"
       ) + 
  scale_color_discrete(labels = c("inches / 10", "inches"))

#static plot
p_daily

#ggplotly(p_daily)

#Nina's code for calculating daily precip sums 
# precipDaily <- precip15 %>%
#   mutate(DATE = lubridate::floor_date(TIMESTAMP, unit = "day")) %>%
#   group_by(DATE) %>%
#   dplyr::summarize(South = sum(South)) %>%
#   ungroup()
```

# Air Temperature Every 30 Min

-   Read in the data from the following Box file path:
    `External-MEF_DATA/EDI/airT_RH_BP_30min/edi.990.1/data_objects/`

-   Clean the data into a tidy format

    -   Subset precipitation data to 2017 - present

-   Plot

```{r airtemp_30}
#set file path to read in data from Box 
filepath = "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/EDI/airT_RH_BP_30min/edi.990.1/data_objects/"

#read in the raw data 
air_temp_30 <- read_csv(here(filepath, "MEF_S2Bog_Met_30min.csv"))

#clean the data 
air_temp_30_clean <- air_temp_30 %>% 
  clean_names() %>% 
  rename(airtempc = air_t,
         rel_humidity = rh,
         datetime = timestamp) %>% 
  mutate(year = format(as.POSIXct(datetime, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y'),
         year = as.numeric(year),
         airtempc_10 = (airtempc / 10),
         airtempc_100 = (airtempc / 100),
         #create a variable that indicates if the airtemp is above or below freezing
         freezing = 0
         ) %>% 
  pivot_longer(
    cols = c("airtempc", "airtempc_10", "airtempc_100"),
    names_to = "airtemp_type",
    values_to = "airtempC"
  )

#assign a value of 1 to the variable if airtemp is above freezing 
air_temp_30_clean$freezing[air_temp_30_clean$airtempC > 0] <- 1

#save clean data CSV to use in future RMD files
write.csv(x = air_temp_30_clean, 
          file = file.path(here("intermediate_data", "air_temp_30_clean.csv")),
          row.names = FALSE)

#plot 
p_air_30 <- ggplot(data = air_temp_30_clean) + 
  geom_line(aes(x = datetime, 
                y = airtempC, 
                color = airtemp_type)) + 
  theme_classic() + 
  labs(x = "Time", 
       y = "Air Temperature (ºC)", 
       title = paste0("Air Temperature Every 30 Min (", min_year, "-", max_year, ")"), 
       color = "Air Temperature Type"
       ) + 
  theme(plot.title = element_text(hjust = 0.5),
          plot.subtitle = element_text(hjust = 0.5)) + 
  scale_color_discrete(labels = c("ºC", 
                                  "ºC / 10", 
                                  "ºC / 100"))

#static plot
p_air_30

#interactive plot
#ggplotly(p_air_30)
```

# Peatland Daily Water Table (Bogwell Data)

More information about peatland and upland water elevation data
collected at the Marcell Experimental Forest can be found on the
[Environmental Data Initiative Repository
Portal](https://data.nal.usda.gov/dataset/marcell-experimental-forest-peatland-and-upland-water-table-elevations).

## Manual Stripcharts

-   Read in the most up-to-date data from the following Box file path:
    `External-MEF_DATA/Hydro/WaterTableElevation/Bogwells/L1`

    -   Less up-to-date, already published data can be found at the
        following Box file path:
        `External-MEF_DATA/EDI/peatland_water_table_daily/edi.562.2/data_objects/`

-   Clean the data into a tidy format

    -   Subset data to 2017 - present

```{r bogwell_clean}
#set file path to read in data from Box 
filepath = "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/Hydro/WaterTableElevation/Bogwells/L1"

#read in the raw data 
bogwell <- read_csv(here(filepath, "L1_Peatland_well_daily_history_2021forMia.csv"))

#clean the data
##note: we do not want any E values for the flag column 
bogwell_clean <- bogwell %>% 
  clean_names() %>% 
  mutate(year = format(as.POSIXct(date, format = '%Y-%m-%d', 
                                  tz = "GMT"), 
                       format = '%Y'),
         year = as.numeric(year)) %>% 
  subset(year >= min_year & year <= max_year)

#save clean data CSV to use in future RMD files
write.csv(x = bogwell_clean, 
          file = file.path(here("intermediate_data", "bogwell_clean.csv")),
          row.names = FALSE)
```

```{r bogwell_plot, eval=FALSE, include=FALSE}
#plot
p_bogwell <- ggplot(data = bogwell_clean) + 
  geom_line(aes(x = date, y = wte, color = peatland)) + 
  theme_classic() + 
  labs(x = "Time", 
       y = "Water Table Elevation (m above sea level)",
       title = paste0("Daily Peatland Water Table Elevation (", min_year, "-", max_year, ")") 
       ) + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7))

#static plot 
p_bogwell
```

### S2 Manual Stripchart

```{r bogwell_s2}
bogwell_clean_s2 <- bogwell_clean %>% 
    subset(peatland == "S2")

#plot
p_bogwell_s2 <- ggplot(data = bogwell_clean_s2) + 
  geom_line(aes(x = date, 
                y = wte, 
                group = 1,
                text = paste0("Date: ", date, "\n",
                              "Water Table Elevation (m): ", wte))) + 
  theme_classic() + 
  labs(x = "Time", 
       y = "Water Table Elevation (m above sea level)",
       title = paste0("S2 Daily Peatland Water Table Elevation (", min_year, "-", max_year, ")") 
       ) + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7))

#interactive plot 
ggplotly(p_bogwell_s2,
         tooltip = "text")
```

Statistics

```{r bogwell_s2_stats}
bogwell_clean_s2 %>% 
  #remove rows that have NA wte values 
  filter(!is.na(wte)) %>% 
  summarise(mean_wte_m = round(mean(wte), digits = 3),
            max_wte_m = max(wte),
            min_wte_m = min(wte), 
            sd_wte_m = round(sd(wte), digits = 3)
            ) %>% 
  #t() %>% 
  kable(caption = "S2 Stripchart Bogwell Statistics", 
        col.names = c("Mean WTE (m)", 
                      "Max WTE (m)",
                      "Min WTE (m)",
                      "SD")) %>% 
  kable_material(c("striped", "hover"))
```

## Shaft Encoders

Function for loading Campbell DataLogger .dat files:

```{r dat_function}
readCDL = function(file){
  
  # read data file starting on 5th line
  dat <- read.csv(file, sep=",",header=FALSE,skip=4,na.strings="NAN",stringsAsFactors=F)
  
  # Read in just the header line (l2)
  # unlist the line, and remove quotes 
  h <- readLines(file, n=2)[2]
  n <- as.factor(unlist(strsplit(h, ",")) )
  n2 <- gsub('"', "", n)
  
  # assign column names to dataframe
  colnames(dat) = n2
  
  return(dat)
}
```

Constants to convert between feet and meters:

```{r constants}
MetersPerFoot = 0.3048
FeetPerMeter = 3.28048

min_year = 2020
max_year = 2021
```

-   Read in the data from the following Box file path:
    `External-MEF_DATA/DataDump/RealTimeData`

-   Clean the data into a tidy format

    -   Subset data to 2020 - present

```{r eval=FALSE, include=FALSE}
#set file path to read in data from Box 
filepath = "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/DataDump/RealTimeData"

#read in the raw data 
s1_se <- readCDL(file = here(filepath, "S1-EM3_Table1.dat"))

#clean the data 
s1_se_clean <- s1_se %>% 
  clean_names() %>% 
  #select relevant columns
  select(timestamp,
         wt_elev) %>% 
  #rename columns 
  rename(datetime = timestamp,
         wt_elev_ft = wt_elev) %>% 
  #reformat columns into appropriate classes 
  mutate(datetime = as.POSIXct(datetime, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"),
    wt_elev_ft = as.numeric(wt_elev_ft),
    wt_elev_m = wt_elev_ft * MetersPerFoot)
```

```{r s1_append, eval=FALSE, include=FALSE}
#set file path to read in data from Box 
filepath = "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/DataDump/RealTimeData"

#read in appended data 

```

### S2 Shaft Encoder

-   We are interested in the `MaxWTElev` column

```{r s2_dat}
#set file path to read in data from Box 
filepath = "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/DataDump/RealTimeData"

#read in the raw data 
s2_se <- readCDL(file = here(filepath, "S2_bogwell_S2BW_Daily.dat"))

s2_se_clean <- s2_se %>% 
  clean_names() %>% 
  #select relevant columns
  select(timestamp,
         max_wt_elev) %>% 
  #rename columns 
  rename(datetime = timestamp,
         max_wt_elev_ft = max_wt_elev) %>% 
  #reformat columns into appropriate classes 
  mutate(datetime = as.POSIXct(datetime, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"),
    max_wt_elev_ft = as.numeric(max_wt_elev_ft),
    max_wt_elev_m = max_wt_elev_ft * MetersPerFoot)
```

-   Plot

```{r s2_dat_plot}
#plot 
p_s2 <- ggplot(data = s2_se_clean) + 
  geom_line(aes(x = datetime, 
                y = max_wt_elev_m,
                group = 1,
                text = paste0("Date: ", datetime, "\n",
                              "Water Table Elevation (m): ", max_wt_elev_m))) + 
  theme_classic() + 
  labs(x = "Time", 
       y = "Water Table Elevation (m)", 
       title = paste0("S2 Bogwell Water Table Elevation (", min_year, "-", max_year, ")")
       ) + 
  theme(plot.title = element_text(hjust = 0.5),
          plot.subtitle = element_text(hjust = 0.5))

#interactive plot
ggplotly(p_s2,
         tooltip = "text")
```

```{r s2_dat_stats}
s2_se_clean %>% 
  #rename columns 
  rename(wte = max_wt_elev_m) %>% 
  #remove rows that have NA wte values 
  filter(!is.na(wte)) %>% 
  summarise(mean_wte_m = round(mean(wte), digits = 3),
            max_wte_m = max(wte),
            min_wte_m = min(wte), 
            sd_wte_m = round(sd(wte), digits = 3)
            ) %>% 
  #t() %>% 
  kable(caption = "S2 Shaft Encoder Bogwell Statistics", 
        col.names = c("Mean WTE (m)", 
                      "Max WTE (m)",
                      "Min WTE (m)",
                      "SD")) %>% 
  kable_material(c("striped", "hover"))
```







```{r bogwell_daily, eval=FALSE, include=FALSE}
#set file path to read in data from Box 
filepath = "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/EDI/peatland_water_table_daily/edi.562.2/data_objects/"

#read in the raw data 
bogwell <- read_csv(here(filepath, "MEF_daily_peatland_water_table.csv"))

#clean the data
bogwell_clean <- bogwell %>% 
  clean_names() %>% 
  subset(peatland == "S2") %>% 
  mutate(year = format(as.POSIXct(date, format = '%Y-%m-%d', 
                                  tz = "GMT"), 
                       format = '%Y'),
         year = as.numeric(year)) %>% 
  subset(year >= min_year & year <= max_year)

#save clean data CSV to use in future RMD files
write.csv(x = bogwell_clean, 
          file = file.path(here("intermediate_data", "bogwell_clean.csv")),
          row.names = FALSE)
```

```{r bogwell_daily_plot, eval=FALSE, include=FALSE}
#plot
p_bogwell <- ggplot(data = bogwell_clean) + 
  geom_line(aes(x = date, y = wte)) + 
  theme_classic() + 
  labs(x = "Time", 
       y = "Water Table Elevation (m above sea level)",
       title = paste0("S2 Daily Peatland Water Table Elevation (", min_year, "-", max_year, ")") 
       )

#static plot 
p_bogwell

#interactive plot
#ggplotly(p_bogwell)
```

Note that this file is being knit as
[`index.html`](https://github.com/miaforsline/climate_bogwell/blob/main/index.html)
into the [`climate_bogwell`
repository](https://github.com/miaforsline/climate_bogwell) in order to
update the [GitHub pages
website](https://miaforsline.github.io/climate_bogwell/), where the
plots can be viewed online.
