==> 4_comparisons.Rmd <==
---
title: 'Marcell Experimental Forest: S2 Stream Height Comparison Analyses (2017 - 2019)'
author: "Mia Forsline"
date: '`r format(Sys.time(), "%B %d, %Y")`'
output:
  html_document:
    theme: flatly
    code_folding: show
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
editor_options:
  markdown:
    wrap: 72
---

# Set Up

-   Install/load necessary R packages
-   Set working directory if necessary (or create a file path to use
    throughout the RMD to call the data from Box)
-   Set constant values for reused variables

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      message = FALSE, 
                      warning = FALSE)

#install packages if necessary, then load libraries
if (!require(librarian)){
  install.packages("librarian")
  library(librarian)
}

librarian::shelf(
  here, #for file paths 
  janitor, 
  plotly, #for creating interactive plots 
  readxl, 
  tidyverse,
  zoo #for interpolating missing data points 
  )

#read in necessary CSV files from prior RMDs
all_streamflow <- read_csv(here("intermediate_data", "all_streamflow.csv"))
air_temp_30_clean <- read_csv(here("intermediate_data", "air_temp_30_clean.csv"))
bogwell_clean <- read_csv(here("intermediate_data", "bogwell_clean.csv")) 
mc_clean <- read_csv(here("intermediate_data", "mc_clean.csv")) %>% 
  rename(datetime = collected)
precip_15_clean <- read_csv(here("intermediate_data", "precip_15_clean.csv")) 
precip_6h_clean <- read_csv(here("intermediate_data", "precip_6h_clean.csv")) 
precip_12h_clean <- read_csv(here("intermediate_data", "precip_12h_clean.csv")) 
precip_daily_clean <- read_csv(here("intermediate_data", "precip_daily_clean.csv")) 
se_clean <- read_csv(here("intermediate_data", "se_clean.csv")) 

# Set working directory path
## Mia's working directory path
#setwd("/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/Hydro/Streamflow/L0_subdaily/StripCharts/AnnualBreakPoint/S2")
## Nina's working directory path 
#setwd("~/../../Box/External-MEF_DATA/Hydro/Streamflow/L0_subdaily/StripCharts")

#set constants 
min_year = 2017
max_year = 2019
```

# Comparisons

## Streamflow + Cumulative Precipitation + Manual Checkpoints

### Cumulative Precipitation Every 6 Hours

```{r p_6h}
#calculate the ratio of the 2 y-axes to plot the variables together
trans_value <- max(precip_6h_clean$precip_10_in, na.rm = TRUE) / max(all_streamflow$stream_height_ft, na.rm = TRUE)

#plot 
p_6h <- ggplot() + 
  #precipitation data 
  geom_col(data = precip_6h_clean, 
            aes(x = datetime, 
                y = precip_10_in,
                colour = I("grey")),
            size = 0.25) + 
  #streamflow data 
  geom_line(data = all_streamflow,
             aes(x = datetime, 
                 y = stream_height_ft),
            color = "blue",
            size = 0.4) + 
  #manual check points
  geom_point(data = mc_clean, 
             aes(x = datetime, y = stripchart_stage),
             size = 0.5) + 
  theme_classic() + 
  labs(x = "Time", 
       title = "2017 - 2019 S2 Bog Streamflow and Cumulative Precipitation Every 6 Hours",
       subtitle = "Streamflow is plotted in blue. \n Precipitation is plotted in grey. \n Manual streamflow checks are plotted as black points.") + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7)) +
  scale_y_continuous(
    #features of the first axis
    name = "Stream Height (ft)",
    #add a second axis and specify its features
    ##determine the ratio between the max value of Axis 1 vs the max value of Axis 2 to decide what number to put in the trans argument
    sec.axis = sec_axis(trans = ~.*trans_value, name = "Cumulative Precipitation / 10 (in)")
  )

#static plot
p_6h

#save the plot in the figures folder 
ggsave(filename = "streamflow_precip_6h_mc.png",
       plot = p_6h,
       path = "figures/",
       width = 8,
       height = 4,
       units = c("in"),
       dpi = 300)

#interactive plot 
ggplotly(p_6h)

```

### Cumulative Precipitation Every 12 Hours

```{r p_12h}
#calculate the ratio of the 2 y-axes to plot the variables together
trans_value <- max(precip_12h_clean$precip_10_in, na.rm = TRUE) / max(all_streamflow$stream_height_ft, na.rm = TRUE)

#plot 
p_12h <- ggplot() + 
  #precipitation data 
  geom_col(data = precip_12h_clean, 
            aes(x = datetime, 
                y = precip_10_in,
                colour = I("grey")),
            size = 0.25) + 
  #streamflow data 
  geom_line(data = all_streamflow,
             aes(x = datetime, 
                 y = stream_height_ft),
            color = "blue",
            size = 0.4) + 
  #manual check points
  geom_point(data = mc_clean, 
             aes(x = datetime, y = stripchart_stage),
             size = 0.5) + 
  theme_classic() + 
  labs(x = "Time", 
       title = "2017 - 2019 S2 Bog Streamflow and Cumulative Precipitation Every 12 Hours",
       subtitle = "Streamflow is plotted in blue. \n Precipitation is plotted in grey. \n Manual streamflow checks are plotted as black points.") + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7)) +
  scale_y_continuous(
    #features of the first axis
    name = "Stream Height (ft)",
    #add a second axis and specify its features
    ##determine the ratio between the max value of Axis 1 vs the max value of Axis 2 to decide what number to put in the trans argument
    ###for example, max value of precipitation / max value of streamflow = 0.533...
    sec.axis = sec_axis(trans = ~.*trans_value, name = "Cumulative Precipitation / 10 (in)")
  )

#static plot
p_12h

#save the plot in the figures folder 
ggsave(filename = "streamflow_precip_12h_mc.png",
       plot = p_12h,
       path = "figures/",
       width = 8,
       height = 4,
       units = c("in"),
       dpi = 300)

#interactive plot 
ggplotly(p_12h)

```

### Cumulative Daily Precipitation

```{r p_daily}
#format precipitation data for a smooth joining process
precip_daily_join <- precip_daily_clean %>% 
  pivot_wider(
    names_from = "precip_type",
    values_from = "precip_in"
  ) %>% 
  subset(year >= min_year & year <= max_year) %>% 
  rename(precip_10_in = precip_10) %>% 
  mutate(datetime = format(as.POSIXct(date, format = '%Y-%m-%d', 
                                  tz = "GMT"),
                           format = '%Y-%m-%d %H:%M:%S'),
         datetime = as.POSIXct(datetime))

#join streamflow and daily precipitation data 
streamflow_precip_daily <- full_join(x = all_streamflow, 
                                     y = precip_daily_join, 
                                     by = "date") %>% 
  select(-precip_in) %>% 
  subset(year >= min_year & year <= max_year)

#calculate the ratio of the 2 y-axes to plot the variables together
trans_value <- max(precip_daily_join$precip_10_in, na.rm = TRUE) / max(all_streamflow$stream_height_ft, na.rm = TRUE)

#plot 
p_daily <- ggplot() + 
  #precipitation data 
  geom_col(data = precip_daily_join, 
            aes(x = datetime, 
                y = precip_10_in,
                colour = I("grey")),
            size = 0.25) + 
  #streamflow data 
  geom_line(data = all_streamflow,
             aes(x = datetime, 
                 y = stream_height_ft),
            color = "blue",
            size = 0.4) + 
  #manual check points
  geom_point(data = mc_clean, 
             aes(x = datetime, y = stripchart_stage),
             size = 0.5) + 
  theme_classic() + 
  labs(x = "Time", 
       title = "2017 - 2019 S2 Bog Streamflow and Daily Cumulative Precipitation",
       subtitle = "Streamflow is plotted in blue. \n Precipitation is plotted in grey. \n Manual streamflow checks are plotted as black points.") + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7)) +
  scale_y_continuous(
    #features of the first axis
    name = "Stream Height (ft)",
    #add a second axis and specify its features
    ##determine the ratio between the max value of Axis 1 vs the max value of Axis 2 to decide what number to put in the trans argument
    ###for example, max value of precipitation / max value of streamflow = 0.533...
    sec.axis = sec_axis(trans = ~.*trans_value, name = "Cumulative Daily Precipitation / 10 (in)")
  )

#static plot
p_daily

#save the plot in the figures folder 
ggsave(filename = "streamflow_precip_daily_mc.png",
       plot = p_daily,
       path = "figures/",
       width = 8,
       height = 4,
       units = c("in"),
       dpi = 300)

#interactive plot 
#ggplotly(p_daily)
```

## Streamflow + air temperature data (every 30min) + manual checks

```{r p_air}
#format air temperature data for plotting 
airtemp_join <- air_temp_30_clean %>% 
  pivot_wider(
    names_from = "airtemp_type",
    values_from = "airtempC"
  ) %>% 
  subset(year >= min_year & year <= max_year) %>% 
  mutate(freezing = as.factor(freezing))

#calculate the ratio of the 2 y-axes to plot the variables together
trans_value <- max(airtemp_join$airtempc_100, na.rm = TRUE) / max(all_streamflow$stream_height_ft, na.rm = TRUE)

#plot 
p_air <- ggplot() + 
  #airtemp data
  geom_line(data = airtemp_join, 
            aes(x = datetime,
                y = airtempc_100),
            color = "pink", 
            alpha = 0.5) +
  #streamflow data 
  geom_line(data = all_streamflow,
             aes(x = datetime, 
                 y = stream_height_ft),
            color = "blue",
            size = 0.75) + 
  #manual check points
  geom_point(data = mc_clean, 
             aes(x = datetime, 
                 y = stripchart_stage),
             size = 0.5) + 
  theme_classic() + 
  labs(x = "Time", 
       title = "S2 Bog Streamflow and Air Temperature (2017 - 2019)",
       subtitle = "Air temperature is plotted in pink. \n Streamflow is plotted in blue. \n Manual streamflow checkpoints are plotted as black dots.") + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7)) + 
  scale_y_continuous(
    # Features of the first axis
    name = "Air Temperature / 100 (ºC)",
    # Add a second axis and specify its features
    sec.axis = sec_axis(trans = ~.*trans_value, name = "Stream Height (ft)")
  ) 

#static plot
p_air

#save the plot in the figures folder 
ggsave(filename = "streamflow_airtemp_manual_checks.png",
       plot = p_air,
       path = "figures/",
       width = 8,
       height = 4,
       units = c("in"),
       dpi = 300)

#interactive plot
#ggplotly(p_air)
```

## Precip (15min) + air temp (30min)

```{r}
precip_15_join <- precip_15_clean %>% 
  pivot_wider(
    names_from = "precip_type",
    values_from = "precip_in"
  ) %>% 
  subset(year >= min_year & year <= max_year)

precip_air <- full_join(x = precip_15_join, 
                        y = airtemp_join, 
                        by = c("datetime", "year")) %>% 
  subset(year >= min_year & year <= max_year)

p_pa <- ggplot() + 
  #airtemp data
  geom_line(data = airtemp_join, 
            aes(x = datetime, 
                y = airtempc_100),
            color = "pink",
            alpha = 0.5,
            size = 0.25) + 
  #precipitation data 
  geom_line(data = precip_15_join, 
            aes(x = datetime, 
                y = precip_10,
                colour = I("grey")),
            size = 0.25) +
  theme_classic() + 
  labs(x = "Time", 
       title = "S2 Precipitation Every 15min and Air Temperature Every 30min",
       subtitle = "Cumulative precipitation is plotted in grey. \n Air temperature is plotted in pink.") + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7)) + 
  scale_y_continuous(
    # Features of the first axis
    name = "Air Temperature (ºC) / 100",
    # Add a second axis and specify its features
    sec.axis = sec_axis(trans = ~., name = "Cumulative Precipitation (in) / 10")
  ) 

#static plot
p_pa

ggsave(filename = "precip_15_airtemp_30.png",
       plot = p_pa,
       path = "figures/",
       width = 10,
       height = 5,
       units = c("in"),
       dpi = 300)

#interactive plot
#ggplotly(p_pa)
```

## Streamflow + bogwell data

```{r}
#calculate median + minimum bogwell water table elevation (wte) values 
median_bog <- median(bogwell_clean$wte)
min_bog <- min(bogwell_clean$wte)

#calculate the difference of wte values from the median and min 
bogwell_diff <- bogwell_clean %>% 
  mutate(diff_from_med = bogwell_clean$wte - median_bog,
         diff_from_min = bogwell_clean$wte - min_bog)

#sum up streamflow by day 
streamflow_daily <- all_streamflow %>% 
  group_by(date) %>% 
  summarise(sum_stream_height_ft = sum(stream_height_ft, na.rm = TRUE))

#combine daily precipitation + daily streamflow sums 
streamflow_precip_daily <- full_join(x = precip_daily_clean,
                                     y = streamflow_daily,
                                     by = "date") %>% 
  pivot_wider(names_from = "precip_type",
              values_from = "precip_in") %>% 
  select(-south_flag) %>% 
  subset(year <= max_year)

#combine daily precipitation/streamflow with bogwell data 
streamflow_precip_bogwell <- full_join(x = streamflow_precip_daily, y = bogwell_diff, by = c("date", "year")) %>% 
  select(-flag,
         -peatland) %>% 
  mutate(wte_100 = (wte / 100)) %>% 
  subset(year <= max_year)


#plot 
p_sb <- ggplot() + 
 # #precipitation data 
 #  geom_line(data = streamflow_precip_bogwell, 
 #            aes(x = date, 
 #                y = precip_10,
 #                colour = I("grey")),
 #            size = 0.25) + 
  #precipitation data 
  geom_line(data = streamflow_precip_bogwell,
             aes(x = date, 
                 y = precip_10),
            color = "blue",
            size = 0.25,
            alpha = 0.5) + 
  #bogwell data
  geom_bar(data = streamflow_precip_bogwell,
            aes(x = date,
                y = diff_from_min)
           ) +
  theme_classic() + 
  labs(x = "Time", 
       title = "S2 Bog Streamflow and Daily Peatland Water Table Elevation",
       subtitle = "Precip is plotted in blue. Bogwell data is plotted in black.") + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7)) + 
  scale_y_continuous(
    # Features of the first axis
    name = "Peatland WTE / 100 (m above sea level)",
    # Add a second axis and specify its features
    sec.axis = sec_axis(trans = ~.*1, name = "Stream Height (ft)")
  ) 

#can scale the y-axis so cut off the top of the precip 

#static plot
#p_sb

#save the plot in the figures folder 
# ggsave(filename = "streamflow_bogwell.png",
#        plot = p_sb,
#        path = "figures/",
#        width = 6,
#        height = 3,
#        units = c("in"),
#        dpi = 300)

#interactive plot 
#ggplotly(p_sb)
```



### Streamflow + precipitation (every 15min)

```{r}
#format precipitation data for a smooth joining process
precip_15_join <- precip_15_clean %>% 
  pivot_wider(
    names_from = "precip_type",
    values_from = "precip_in"
  ) %>% 
  subset(year >= min_year & year <= max_year)

#join streamflow and precipitation data 
streamflow_precip_15 <- full_join(x = all_streamflow, y = precip_15_join, by = "datetime") %>% 
  rename(precip_10_in = precip_10) %>% 
  select(-precip_in) 

#plot 
p_sp <- ggplot() + 
  #precipitation data 
  geom_line(data = streamflow_precip_15, 
            aes(x = datetime, 
                y = precip_10_in,
                colour = I("grey")),
            size = 0.25) + 
  #streamflow data 
  geom_line(data = all_streamflow,
             aes(x = datetime, 
                 y = stream_height_ft),
            color = "blue",
            size = 0.25) + 
  theme_classic() + 
  labs(x = "Time", 
       title = "2017 - 2019 S2 Bog Streamflow and Scaled Precipitation",
       subtitle = "Streamflow is plotted in blue. Precipitation is plotted in black.") + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7)) + 
  scale_y_continuous(
    # Features of the first axis
    name = "Stream Height (ft)",
    # Add a second axis and specify its features
    sec.axis = sec_axis(trans = ~.*0.698509267, name = "Precipitation Sum Scaled by 10 (in)")
  ) 

#static plot
#p_sp

#save the plot in the figures folder 
ggsave(filename = "streamflow_precip.png",
       plot = p_sp,
       path = "figures/",
       width = 6,
       height = 3,
       units = c("in"),
       dpi = 300)

#interactive plot
#ggplotly(p_sp)
```

### Stripchart + old shaft encoder data

```{r}
stripchart_se <- full_join(x = all_streamflow, y = se_clean, by = "datetime")

p_stripchart_se <- ggplot() + 
  geom_line(data = all_streamflow, 
            aes(x = datetime, y = stream_height_ft),
            color = "blue",
            size = 0.25) + 
  geom_line(data = se_clean, 
            aes(x = datetime, y = stage),
            color = "red",
            alpha = 0.75,
            size = 0.25) + 
  theme_classic() + 
  labs(x = "Time", 
       y = "Stream Height (ft)",
       title = "Stripchart vs Shaft Encoder S2 Streamflow (2017 - 2019)",
       subtitle = "Stripchart data is in blue. Shaft encoder data is in red.") + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7))

#static plot
#p_stripchart_se

ggsave(filename = "stripchart_shaft_encoder.png",
       plot = p_stripchart_se,
       path = "figures/",
       width = 6,
       height = 3,
       units = c("in"),
       dpi = 300)

ggplotly(p_stripchart_se)
```
