==> 5_air_comparisons.Rmd <==
---
title: 'Marcell Experimental Forest: S2 Stream Height & Air Temperature Comparison Analyses'
author: "Mia Forsline"
date: '`r format(Sys.time(), "%B %d, %Y")`'
knit: (function(inputFile, encoding) { 
      rmarkdown::render(inputFile,
                        encoding=encoding, output_file="/Users/miaforsline/Desktop/air_comparisons/index.html") })
output:
  html_document:
    theme: flatly
    code_folding: hide
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
editor_options:
  markdown:
    wrap: 72
---

# Objectives

The goal of this step in the workflow is to compare S2 streamflow height
to air temperature data in order to quality check the streamflow data.

# Set Up

-   Install/load necessary R packages

-   Set working directory if necessary (or create a file path to use
    throughout the RMD to call the data from Box)

```{r setup, include=FALSE}
knitr::opts_chunk$set(
	echo = TRUE,
	message = FALSE,
	warning = FALSE
)
#install packages if necessary, then load libraries
if (!require(librarian)){
  install.packages("librarian")
  library(librarian)
}

librarian::shelf(
  here, #for file paths 
  janitor, 
  plotly, #for creating interactive plots 
  readxl, 
  tidyverse,
  zoo #for interpolating missing data points 
  )
```

-   Set date range of interest

```{r min_max_year}
#set constants 
min_year = 2017
max_year = 2021
```

-   Recreate intermediate data for later use

```{r intermediate_data, message=FALSE, warning=FALSE}
#2017 
filepath <- "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/Hydro/Streamflow/L0_subdaily/StripCharts/AnnualBreakPoint/S2"

#read in the 2017 data 
wy2017 <- read_excel(path = here(filepath, "WY2017.S2_Breakpoint.xlsx"), 
                     skip =3) #skip the top 3 rows of the Excel Sheet 

#clean the 2017 data
wy2017_clean <- wy2017 %>% 
  #change column names
  rename(datetime = "Date/time",
         stream_height_ft = "Stage.ft") %>% 
  #remove unnecessary columns 
  select(c("datetime", "stream_height_ft")) %>% 
  #format column names into lower_snake_case
  clean_names() %>% 
  #create a date column 
  mutate(date = format(as.POSIXct(datetime, format = '%m/%d/%Y %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y-%m-%d'),
         date = as.POSIXct(date, tz = "GMT")) 

#2018

filepath <- "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/Hydro/Streamflow/L0_subdaily/StripCharts/AnnualBreakPoint/S2"

#create a function to specify custom time intervals 
`%between%` <- function(x, interval) x >= interval[1] & x <= interval[2]

#read in the 2018 data 
wy2018 <- read_excel(path = here(filepath, "WY2018.S2_Breakpoint.xlsx"), 
                     skip =3)

#clean the 2018 data
wy2018_clean <- wy2018 %>% 
  rename(datetime = "Date/time",
         stream_height_ft = "Stage.ft") %>% 
  select(c("datetime", "stream_height_ft")) %>% 
  clean_names() %>% 
  mutate(date = format(as.POSIXct(datetime, format = '%m/%d/%Y %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y-%m-%d'),
         date = as.POSIXct(date, tz = "GMT")) 

#assign October NA values 
wy2018_clean$stream_height_ft[wy2018_clean$date %between% as.Date(c('2018-10-09', '2018-10-10'))] <- NA
wy2018_clean$stream_height_ft[wy2018_clean$date %between% as.Date(c('2018-10-11', '2018-10-15'))] <- NA

#assign November NA values 
wy2018_clean$stream_height_ft[wy2018_clean$date %between% as.Date(c('2018-11-10', '2018-11-14'))] <- NA
wy2018_clean$stream_height_ft[wy2018_clean$date %between% as.Date(c('2018-11-25', '2018-12-04'))] <- NA

#assign December NA values 
wy2018_clean$stream_height_ft[wy2018_clean$date %between% as.Date(c('2018-12-01', '2018-12-03'))] <- NA
wy2018_clean$stream_height_ft[wy2018_clean$date %between% as.Date(c('2018-12-08', '2018-12-10'))] <- NA
wy2018_clean$stream_height_ft[wy2018_clean$date %between% as.Date(c('2018-12-11', '2018-12-31'))] <- NA

wy2018_altered <- wy2018_clean
wy2018_altered <- subset(wy2018_altered, !(wy2018_altered$date %between% as.Date(c('2018-10-31', '2018-11-01')) & wy2018_altered$stream_height_ft < 0.10))

filepath <- "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/Hydro/Streamflow/L0_subdaily/StripCharts/AnnualBreakPoint/S2"

#2019 
#read in the 2019 data 
wy2019 <- read_excel(path = here(filepath, "WY2019.S2_Breakpoint.xlsx"), 
                     skip =3)

#clean the 2019 data
wy2019_clean <- wy2019 %>% 
  rename(datetime = "Date/time",
         stream_height_ft = "Stage.ft") %>% 
  select(c("datetime", "stream_height_ft")) %>% 
  clean_names() %>% 
  mutate(date = format(as.POSIXct(datetime, format = '%m/%d/%Y %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y-%m-%d'),
         date = as.POSIXct(date, tz = "GMT")) 

#assign February NA values 
wy2019_clean$stream_height_ft[wy2019_clean$date %between% as.Date(c('2019-04-01', '2019-04-02'))] <- NA

#assign March NA values 
wy2019_clean$stream_height_ft[wy2019_clean$date == as.Date('2019-03-14')] <- NA
wy2019_clean$stream_height_ft[wy2019_clean$date == as.Date('2019-03-18')] <- NA
wy2019_clean$stream_height_ft[wy2019_clean$date %between% as.Date(c('2019-03-30', '2019-03-31'))] <- NA

#assign April NA values 
wy2019_clean$stream_height_ft[wy2019_clean$date %between% as.Date(c('2019-04-01', '2019-04-02'))] <- NA
wy2019_clean$stream_height_ft[wy2019_clean$date %between% as.Date(c('2019-04-08', '2019-04-17'))] <- NA
wy2019_clean$stream_height_ft[wy2019_clean$date %between% as.Date(c('2019-04-17', '2019-04-22'))] <- NA

#assign September NA values 
wy2019_clean$stream_height_ft[wy2019_clean$date %between% as.Date(c('2019-09-03', '2019-09-10'))] <- NA

#2020 
filepath <- "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/Hydro/Streamflow/L0_subdaily/StripCharts/AnnualBreakPoint/S2"

#read in the 2020 data 
wy2020 <- read_excel(path = here(filepath, "WY2020.S2_Breakpoint.xlsx"), 
                     skip =3)

#clean the 2020 data
wy2020_clean <- wy2020 %>% 
  rename(datetime = "Date/time",
         stream_height_ft = "Stage.ft") %>% 
  select(c("datetime", "stream_height_ft")) %>% 
  clean_names() %>% 
  mutate(date = format(as.POSIXct(datetime, format = '%m/%d/%Y %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y-%m-%d'),
         date = as.POSIXct(date, tz = "GMT")) 

#create date range to fill in with NA values 
start_date <- as.POSIXct("7/28/2020 13:40:00", 
                         format = '%m/%d/%Y %H:%M:%S',
                         tz = "GMT")
end_date <- as.POSIXct("8/11/2020 00:00:00", 
                         format = '%m/%d/%Y %H:%M:%S',
                         tz = "GMT")

#create a function to specify custom time intervals 
`%between%` <- function(x, interval) x >= interval[1] & x <= interval[2]

#assign February NA values 
wy2020_clean$stream_height_ft[wy2020_clean$datetime %between% c(start_date, end_date)] <- NA

#2021 
filepath <- "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/Hydro/Streamflow/L0_subdaily/StripCharts/AnnualBreakPoint/S2"

#read in the 2021 data 
wy2021 <- read_excel(path = here(filepath, "WY2021.S2_Breakpoint.xlsx"), 
                     skip =3)

#clean the 2021 data
wy2021_clean <- wy2021 %>% 
  rename(datetime = "Date/time",
         stream_height_ft = "Stage.ft") %>% 
  select(c("datetime", "stream_height_ft")) %>% 
  clean_names() %>% 
  mutate(date = format(as.POSIXct(datetime, format = '%m/%d/%Y %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y-%m-%d'),
         date = as.POSIXct(date, tz = "GMT")) 

#identify date intervals that need to be filled w/NAs
date1 <- as.POSIXct("2021-03-04 00:00:00", 
                         format = '%Y-%m-%d %H:%M:%S',
                         tz = "GMT")
date2 <- as.POSIXct("2021-03-09 00:00:00", 
                         format = '%Y-%m-%d %H:%M:%S',
                         tz = "GMT")
date3 <- as.POSIXct("2021-11-11 23:59:00", 
                         format = '%Y-%m-%d %H:%M:%S',
                         tz = "GMT")
date4 <- as.POSIXct("2021-12-29 00:00:00", 
                         format = '%Y-%m-%d %H:%M:%S',
                         tz = "GMT")


#create a function to specify custom time intervals 
`%between%` <- function(x, interval) x >= interval[1] & x <= interval[2]

#assign March NA values
wy2021_clean$stream_height_ft[wy2021_clean$datetime %between% c(date1, date2)] <- NA

#assign Nov-Dec NA values 
wy2021_clean$stream_height_ft[wy2021_clean$datetime %between% c(date3, date4)] <- NA

#all water years 
all_streamflow <- rbind(wy2017_clean, 
                        wy2018_altered, 
                        wy2019_clean,
                        wy2020_clean,
                        wy2021_clean)

#air temp data
#set file path to read in data from Box 
filepath = "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/EDI/airT_RH_BP_30min/edi.990.1/data_objects/"

#read in the raw data 
air_temp_30 <- read_csv(here(filepath, "MEF_S2Bog_Met_30min.csv"))

#clean the data 
air_temp_30_clean <- air_temp_30 %>% 
  clean_names() %>% 
  rename(airtempc = air_t,
         rel_humidity = rh,
         datetime = timestamp) %>% 
  mutate(year = format(as.POSIXct(datetime, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y'),
         year = as.numeric(year),
         airtempc_10 = (airtempc / 10),
         airtempc_100 = (airtempc / 100),
         #create a variable that indicates if the airtemp is above or below freezing
         freezing = 0
         ) %>% 
  pivot_longer(
    cols = c("airtempc", "airtempc_10", "airtempc_100"),
    names_to = "airtemp_type",
    values_to = "airtempC"
  )

#manual checks data 
filepath <- "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/Hydro/Streamflow/L0_subdaily/ManualChecks"

#read in the manual checks data 
mc <- read_csv(here(filepath, "2017-2021_S2Stage.csv"))

#clean the data
mc_clean <- mc %>% 
  clean_names() %>% 
  #remove S2 lagging pool data and keep only the S2 weir data
  subset(name == "S2 WEIR") %>%
  mutate(collected = as.POSIXct(collected, format = '%m/%d/%Y %H:%M', 
                                tz = "GMT"),
         year = format(as.POSIXct(collected, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y'),
         date = format(as.POSIXct(collected, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y-%m-%d'),
         diff = point_gage - stripchart_stage
         ) %>% 
  subset(year >= min_year & year <= max_year) %>% 
  rename(datetime = collected)

#precipitation data 
#set file path to read in data from Box 
filepath = "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/EDI/precipitation_15min/edi.849.2/data_objects/"

#read in the raw data 
precip_15 <- read_csv(here(filepath, "MEF_South_precip_15min.csv"))

#clean the data 
precip_15_clean <- precip_15 %>% 
  clean_names() %>% 
  mutate(year = format(as.POSIXct(timestamp, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y'),
         year = as.numeric(year)) %>% 
  subset(year >= min_year) %>% 
  rename(precip_in = south_pcp) %>% 
  mutate(precip_10 = (precip_in / 10)) %>% 
  pivot_longer(
    cols = c("precip_in", "precip_10"),
    names_to = "precip_type",
    values_to = "precip_in"
  ) %>% 
  rename(datetime = timestamp)

#read in 15min precip data 
precip_6h <- precip_15 

#clean and sum precip by 6 hours 
precip_6h_clean <- precip_6h %>% 
  clean_names() %>% 
  mutate(timestamp = lubridate::floor_date(timestamp, unit = "6 hours")) %>%
    group_by(timestamp) %>%
    dplyr::summarize(south_pcp = sum(south_pcp)) %>%
    ungroup() %>% 
  rename(datetime = timestamp,
         precip_in = south_pcp) %>% 
  mutate(
    year = format(as.POSIXct(datetime, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y'),
         year = as.numeric(year),
     date = format(as.POSIXct(datetime, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y-%m-%d'),
    precip_10_in = precip_in / 10
  ) %>% 
  subset(year >= min_year & year <= max_year)

#read in 15min precip data 
precip_12h <- precip_15 

#clean and sum precip by 6 hours 
precip_12h_clean <- precip_12h %>% 
  clean_names() %>% 
  mutate(timestamp = lubridate::floor_date(timestamp, unit = "12 hours")) %>%
    group_by(timestamp) %>%
    dplyr::summarize(south_pcp = sum(south_pcp)) %>%
    ungroup() %>% 
  rename(datetime = timestamp,
         precip_in = south_pcp) %>% 
  mutate(
    year = format(as.POSIXct(datetime, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y'),
         year = as.numeric(year),
    date = format(as.POSIXct(datetime, format = '%Y-%m-%d %H:%M:%S', 
                                  tz = "GMT"), 
                       format = '%Y-%m-%d'),
    precip_10_in = precip_in / 10
  ) %>% 
  subset(year >= min_year & year <= max_year)

#set file path to read in data from Box 
filepath = "/Users/miaforsline/Library/CloudStorage/Box-Box/External-MEF_DATA/EDI/precipitation_daily/edi.563.3/data_objects/"

#read in the raw data 
precip_daily <- read_csv(here(filepath, "MEF_precipitation_daily.csv"))

#clean the data 
precip_daily_clean <- precip_daily %>% 
  clean_names() %>% 
  select(date, south_pcp, south_flag) %>% 
  mutate(year = format(as.POSIXct(date, format = '%Y-%m-%d', 
                                  tz = "GMT"), 
                       format = '%Y'),
         year = as.numeric(year)) %>% 
  subset(year >= min_year) %>% 
  rename(precip_in = south_pcp) %>% 
  mutate(precip_10 = (precip_in / 10)) %>% 
  pivot_longer(
    cols = c("precip_in", "precip_10"),
    names_to = "precip_type",
    values_to = "precip_in"
  )

precip_daily_join <- precip_daily_clean %>% 
  pivot_wider(
    names_from = "precip_type",
    values_from = "precip_in"
  ) %>% 
  subset(year >= min_year & year <= max_year) %>% 
  rename(precip_10_in = precip_10) %>% 
  mutate(datetime = format(as.POSIXct(date, format = '%Y-%m-%d', 
                                  tz = "GMT"),
                           format = '%Y-%m-%d %H:%M:%S'),
         datetime = as.POSIXct(datetime))
```


# Daily Cumulative Precipitation + Air Temperature Every 30min

-   Format the data

```{r}
#format precipitation data for plotting 
precip_15_join <- precip_15_clean %>% 
  pivot_wider(
    names_from = "precip_type",
    values_from = "precip_in"
  ) %>% 
  subset(year >= min_year & year <= max_year) %>% 
  rename(precip_10_in = precip_10)

precip_daily_join <- precip_daily_join %>% 
  mutate(dummy = "b",
         dummy = as.factor(dummy))

#format air temperature data for plotting 
airtemp_join <- air_temp_30_clean %>% 
  pivot_wider(
    names_from = "airtemp_type",
    values_from = "airtempC"
  ) %>% 
  subset(year >= min_year & year <= max_year) %>% 
  mutate(freezing = as.factor(freezing),
         dummy = "a")

#format air temp data for plotting 
precip_air <- full_join(x = precip_15_join, 
                        y = airtemp_join, 
                        by = c("datetime", "year")) %>% 
  subset(year >= min_year & year <= max_year)

#join air and precipitation data 
precip_air <- full_join(x = precip_15_join, 
                        y = airtemp_join, 
                        by = c("datetime", "year")) %>% 
  subset(year >= min_year & year <= max_year)
```

-   Plot

    -   air temperature (ºC / 100) is plotted as a [pink
        line]{style="color:pink"}, and

    -   cumulative daily precipitation (in / 10) is plotted as grey
        bars.

```{r p_daily}
#plot
p_daily <- ggplot() + 
  #airtemp data
  geom_line(data = airtemp_join, 
            aes(x = datetime, 
                y = airtempc_100,
                color = dummy,
                group = 1,
                #create hovering labels for interactive graph 
                text = paste0("DateTime: ", datetime, "\n",
                              "AirTemp (ºC/100): ", airtempc_100)
                ),
            color = "pink",
            alpha = 0.5,
            size = 0.25) + 
  #precipitation data 
  geom_col(data = precip_daily_join, 
            aes(x = datetime, 
                y = precip_10_in,
                colour = dummy,
                group = 1,
                text = paste0("Date: ", date, "\n",
                              "Daily Cumulative Precipitation (in/10): ", precip_10_in)),
            size = 0.25,
            color = I("grey")) +
  theme_classic() + 
  labs(x = "Time", 
       y = " ", 
       #title = "S2 Daily Cumulative Precipitation \n and Air Temperature Every 30min",
       subtitle = "Daily cumulative precipitation is plotted in grey. \n Air temperature is plotted in pink.") + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7)) 

#interactive plot
ggplotly(p_daily,
         tooltip = "text") %>% 
  layout(title = list(text = paste0(paste0("S2 Daily Cumulative Precipitation \n and Air Temperature Every 30min (", min_year, "-", max_year, ")"),
                                    '<br>',
                                    '<sup>',
                                     paste0("Daily cumulative precipitation is plotted in grey. \n Air temperature is plotted in pink."), 
                                    '</sup>')), 
         font = list(size = 12)
         )
```

# Streamflow + Air Temperature Every 30min + Manual Checkpoints

For interactive plots,

-   air temperature (ºC / 100) is plotted as a [pink
    line]{style="color:pink"},

-   streamflow (ft) is plotted as a [blue line]{style="color:blue"},

-   and streamflow manual checkpoints (ft) are plotted as **black
    dots**.

```{r p_air}
#calculate the ratio of the 2 y-axes to plot the variables together in the static plot
trans_value <- max(airtemp_join$airtempc_100, na.rm = TRUE) / max(all_streamflow$stream_height_ft, na.rm = TRUE)

#plot 
p_air <- ggplot() + 
  #airtemp data
  geom_line(data = airtemp_join, 
            aes(x = datetime,
                y = airtempc_100, 
                group = 1,
                #create hovering labels for interactive graph 
                text = paste0("DateTime: ", datetime, "\n",
                              "AirTemp (ºC/100): ", airtempc_100)
                ),
            color = "pink", 
            alpha = 0.4) +
  #streamflow data 
  geom_line(data = all_streamflow,
             aes(x = datetime, 
                 y = stream_height_ft,
                 group = 1,
                 #create hovering labels for interactive graph 
                 text = paste0("DateTime: ", datetime, "\n",
                              "Stream Height (ft): ", round(stream_height_ft, digits = 3))
                 ),
            color = "blue",
            size = 0.75) + 
  #manual check points
  geom_point(data = mc_clean, 
             aes(x = datetime, 
                 y = stripchart_stage,
                 group = 1,
                 #create hovering labels for interactive graph 
                 text = paste0("DateTime: ", datetime, "\n",
                              "Manual Stream Height Check (ft): ", stripchart_stage)
                 ),
             size = 0.5) + 
  theme_classic() + 
  labs(x = "Time", 
       y = " ",
       title = paste0("S2 Bog Streamflow and Air Temperature (", min_year, "-", max_year, ")"),
       subtitle = "Air temperature is plotted in pink. \n Streamflow is plotted in blue. \n Manual streamflow checkpoints are plotted as black dots.") + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7)) + 
  scale_y_continuous(
    # Features of the first axis
    name = "Air Temperature / 100 (ºC)",
    # Add a second axis and specify its features
    sec.axis = sec_axis(trans = ~.*trans_value, name = "Stream Height (ft)")
  ) 

#interactive plot
ggplotly(p_air,
         tooltip = "text") %>% 
  layout(title = list(text = paste0(paste0("S2 Bog Streamflow and Air Temperature (", min_year, "-", max_year, ")"),
                                    '<br>',
                                    '<sup>',
                                     paste0(" "), 
                                    '</sup>')), 
         font = list(size = 12)
         )
```

# Streamflow + Daily Cumulative Precipitation + Air Temperature Every 30min + Manual Checkpoints

For interactive plots,

-   air temperature (ºC / 100) is plotted as a [pink
    line]{style="color:pink"},

-   cumulative daily precipitation (in / 10) is plotted as grey bars,

-   streamflow (ft) is plotted as a [blue line]{style="color:blue"},

-   and streamflow manual checkpoints (ft) are plotted as **black
    dots**.

```{r p_all}
#plot 
p_all <- ggplot() + 
  #airtemp data
  geom_line(data = airtemp_join, 
            aes(x = datetime,
                y = airtempc_100, 
                group = 1,
                #create hovering labels for interactive graph 
                text = paste0("DateTime: ", datetime, "\n",
                              "AirTemp (ºC/100): ", airtempc_100)
                ), 
            color = "pink",
            alpha = 0.4) +
  #streamflow data 
  geom_line(data = all_streamflow,
             aes(x = datetime, 
                 y = stream_height_ft,
                 group = 1,
                 #create hovering labels for interactive graph 
                 text = paste0("DateTime: ", datetime, "\n",
                              "Stream Height (ft): ", round(stream_height_ft, digits = 3))
                 ),
            color = "blue",
            size = 0.75) + 
  #manual check points
  geom_point(data = mc_clean, 
             aes(x = datetime, 
                 y = stripchart_stage,
                 group = 1,
                 #create hovering labels for interactive graph 
                 text = paste0("DateTime: ", datetime, "\n",
                              "Manual Stream Height Check (ft): ", stripchart_stage)
                 ),
             size = 0.5) + 
  #daily precipitation data 
  geom_col(data = precip_daily_join, 
            aes(x = datetime, 
                y = precip_10_in,
                group = 1,
                text = paste0("Date: ", date, "\n",
                              "Daily Cumulative Precipitation (in/10): ", precip_10_in)),
            colour = I("grey"),
            size = 0.25) +
  theme_classic() + 
  labs(x = "Time", 
       y = " ", 
       title = paste0("S2 Bog Streamflow, Air Temperature, \n and Daily Cumulative Precipitation (", min_year, "-", max_year, ")"),
       subtitle = "Air temperature is plotted in pink and scaled by 100. 
       Precipitation is plotted as grey bars and scaled by 10.
       Streamflow is plotted in blue.
       Manual streamflow checkpoints are plotted as black dots.") + 
  theme(plot.title = element_text(hjust = 0.5),
        plot.subtitle = element_text(hjust = 0.5, 
                                     size = 7),
        legend.position = "none") 

#interactive plot
ggplotly(p_all,
         tooltip = "text") %>% 
  layout(title = list(text = paste0(paste0("S2 Bog Streamflow, Air Temperature, \n and Daily Cumulative Precipitation (", min_year, "-", max_year, ")"),
                                    '<br>',
                                    '<sup>',
                                     paste0(""), 
                                    '</sup>')), 
         font = list(size = 12)
         )
```

Run this code chunk to knit to HTML without using the knit button in
RStudio

-   Save as `index.html` to render GitHub Page

```{r knit_to_html, eval=FALSE, include=FALSE}
#set file names and file path
input_file_name <- "5_air_comparisons.Rmd"
output_file_name <- "index.html"
filepath <- "/Users/miaforsline/Desktop/air_comparisons"

#knit to HTML
rmarkdown::render(input = input_file_name, 
                  output_format = "html_document", 
                  output_file = output_file_name, 
                  output_dir = filepath)
```

# Code Accessibility

Note that this file is being knit as
[`index.html`](https://github.com/miaforsline/air_comparisons/blob/main/index.html)
into the public [`air_comparisons` GitHub
repository](https://github.com/miaforsline/precip_comparisons) in order
to update the [GitHub pages
website](https://miaforsline.github.io/air_comparisons/), where the
plots can be viewed online.
